Added git-fork. Use it eg. to start a new RMG run in parrallel.
Make sure you're in the head (root) of your git repository, then run:
    git-fork branchname

It makes a branch, clones it into a separate folder alongside the current one.
You can then cd into that folder, do your stuff, make commits, and 'git push'
to get the changes into the branch on the original repository.

Useful for running jobs in parrallel and keeping the results in git.